+++ packages/core/lib/src/core_widget_factory.dart
@@ -19,6 +19,7 @@ import 'internal/margin_vertical.dart';
 import 'internal/platform_specific/fallback.dart'
     if (dart.library.io) 'internal/platform_specific/io.dart';
 import 'internal/text_ops.dart' as text_ops;
+import 'utils/roman_numerals_converter.dart';
 
 final _logger = Logger('fwfh.WidgetFactory');
 
@@ -231,12 +232,25 @@ class WidgetFactory extends WidgetFactoryResetter with AnchorWidgetFactory {
     MainAxisAlignment mainAxisAlignment = MainAxisAlignment.start,
     TextBaseline textBaseline = TextBaseline.alphabetic,
   }) {
-    return Flex(
-      crossAxisAlignment: crossAxisAlignment,
-      direction: direction,
-      mainAxisAlignment: mainAxisAlignment,
-      textBaseline: textBaseline,
-      children: children,
+    return LayoutBuilder(
+      builder: (_, bc) {
+        Widget built = Flex(
+          crossAxisAlignment: crossAxisAlignment,
+          direction: direction,
+          mainAxisAlignment: mainAxisAlignment,
+          textBaseline: textBaseline,
+          children: children,
+        );
+        switch (direction) {
+          case Axis.horizontal:
+            built = CssSizingHint(maxWidth: bc.maxWidth, child: built);
+            break;
+          case Axis.vertical:
+            built = CssSizingHint(maxHeight: bc.maxHeight, child: built);
+            break;
+        }
+        return built;
+      },
     );
   }
 
@@ -363,7 +377,7 @@ class WidgetFactory extends WidgetFactoryResetter with AnchorWidgetFactory {
     int index,
   ) {
     final text = getListMarkerText(listStyleType, index);
-    final textStyle = resolved.style;
+    final textStyle = resolved.prepareTextStyle();
     if (text.isNotEmpty) {
       return buildText(tree, resolved, TextSpan(style: textStyle, text: text));
     }
@@ -519,33 +533,15 @@ class WidgetFactory extends WidgetFactoryResetter with AnchorWidgetFactory {
       case kCssListStyleTypeDecimal:
         return '$i.';
       case kCssListStyleTypeRomanLower:
-        final roman = _getListMarkerRoman(i)?.toLowerCase();
+        final roman = intToRomanNumerals(i)?.toLowerCase();
         return roman != null ? '$roman.' : '';
       case kCssListStyleTypeRomanUpper:
-        final roman = _getListMarkerRoman(i);
+        final roman = intToRomanNumerals(i);
         return roman != null ? '$roman.' : '';
       case kCssListStyleTypeNone:
       default:
         return '';
     }
-  }
-
-  String? _getListMarkerRoman(int i) {
-    // TODO: find some lib to generate programatically
-    const map = <int, String>{
-      1: 'I',
-      2: 'II',
-      3: 'III',
-      4: 'IV',
-      5: 'V',
-      6: 'VI',
-      7: 'VII',
-      8: 'VIII',
-      9: 'IX',
-      10: 'X',
-    };
-
-    return map[i];
   }
 
   /// Returns an [AssetImage].
@@ -1138,6 +1134,10 @@ class WidgetFactory extends WidgetFactoryResetter with AnchorWidgetFactory {
         if (whitespace != null) {
           tree.inherit(text_ops.whitespace, whitespace);
         }
+        break;
+
+      case kCssTextShadow:
+        textShadowApply(tree, style);
         break;
     }