+++ packages/core/lib/src/data/normal_line_height.dart
@@ -1,7 +1,0 @@
-part of '../core_data.dart';
-
-class NormalLineHeight {
-  final double? value;
-
-  const NormalLineHeight(this.value);
-}